fix code scanning alerts
