+++ core/agent_harness/turn_context.py
@@ -216,7 +216,6 @@ class TurnContext:
 __all__ = [
     "AgentRuntimeRequest",
     "PromptRenderable",
-    "SystemPromptInput",
     "TurnContext",
     "TurnContextSource",
 ]
