+++ core/execution.py
@@ -9,7 +9,7 @@ from concurrent.futures import Future, ThreadPoolExecutor, as_completed
 from dataclasses import dataclass, field, replace
 from typing import Any
 
-from core.llm.agent_llm_client import ToolCall
+from core.llm.types import ToolCall
 from core.types import AgentTool, AgentToolContext, RuntimeTool
 from platform.observability.tool_trace import redact_sensitive
 from tools.utils.integration_sources import availability_view
