+++ tests/core/agent/test_turn_context_runtime_request.py
@@ -9,7 +9,7 @@ from core.agent import Agent
 from core.agent_harness.conversation_memory import MAX_CONVERSATION_MESSAGES
 from core.agent_harness.prompts import PromptEnvelope
 from core.agent_harness.turn_context import TurnContext
-from core.llm.agent_llm_client import AgentLLMResponse
+from core.llm.types import AgentLLMResponse
 from core.messages import UserRuntimeMessage
 from core.types import AgentTool
