+++ tests/interactive_shell/test_terminal_runtime_turns.py
@@ -8,7 +8,7 @@ import pytest
 from rich.console import Console
 
 from core.agent_harness.session import ReplSession
-from core.llm.agent_llm_client import AgentLLMResponse, ToolCall
+from core.llm.types import AgentLLMResponse, ToolCall
 from interactive_shell.agent_shell.agent import handle_message_with_agent
 from interactive_shell.runtime.core.turn_accounting import (
     ToolCallingTurnResult,
