+++ tools/investigation/stages/gather_evidence/loop.py
@@ -8,7 +8,7 @@ from dataclasses import dataclass
 from typing import Any
 
 from core.context.state.evidence import EvidenceEntry
-from core.llm.agent_llm_client import ToolCall
+from core.llm.types import ToolCall
 from core.llm_invoke_errors import LLMInvokeFailure
 from platform.common.truncation import truncate
